+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - added myems-admin.conf and myems-web.conf for nginx
 - added binding microgrid to space in myems-admin
+- added blank page image for space mnvironment monitor in myems-web
 ### Changed
 - subspace names under the same parent space cannot be duplicated
 - updated power stations